Backtrace: don't add first line if useless
I'm running into a scenario where I'm modifying the stack trace of the
exception being reported. I have the function for the first line, but
php Exceptions don't have a getFunction() to match their getFile() and
getLine(). This change allows including the *whole* stack trace in
getTrace() and having it report correctly (as long as getFile()
returns something falsy).